# Fix computation of shuffle shard when shard size is <= 0, or higher than number of ring instances.