--- a/ring/ring.go
+++ b/ring/ring.go
@@ -775,16 +775,13 @@ func (r *Ring) shuffleShard(identifier string, size int, lookbackPeriod time.Dur
 	// We need to iterate zones always in the same order to guarantee stability.
 	for _, zone := range actualZones {
 		var tokens []uint32
-		var maxAddedInstances int
 
 		if r.cfg.ZoneAwarenessEnabled {
 			tokens = r.ringTokensByZone[zone]
-			maxAddedInstances = r.instancesCountPerZone[zone]
 		} else {
 			// When zone-awareness is disabled, we just iterate over 1 single fake zone
 			// and use all tokens in the ring.
 			tokens = r.ringTokens
-			maxAddedInstances = len(r.ringDesc.Ingesters)
 		}
 
 		// Initialise the random generator used to select instances in the ring.
@@ -793,13 +790,10 @@ func (r *Ring) shuffleShard(identifier string, size int, lookbackPeriod time.Dur
 		// property when the shard size changes or a new zone is added.
 		random := rand.New(rand.NewSource(shardUtil.ShuffleShardSeed(identifier, zone)))
 
-		// If we add maxAddedInstances, we can stop searching for more instances, since there can't be any more of them.
-		addedInstances := 0
-
 		// To select one more instance while guaranteeing the "consistency" property,
 		// we do pick a random value from the generator and resolve uniqueness collisions
 		// (if any) continuing walking the ring.
-		for i := 0; i < numInstancesPerZone && addedInstances < maxAddedInstances; i++ {
+		for i := 0; i < numInstancesPerZone; i++ {
 			start := searchToken(tokens, random.Uint32())
 			iterations := 0
 			found := false
@@ -831,7 +825,6 @@ func (r *Ring) shuffleShard(identifier string, size int, lookbackPeriod time.Dur
 
 				// Include instance in the subring.
 				shard[instanceID] = instance
-				addedInstances++
 
 				// If the lookback is enabled and this instance has been registered within the lookback period
 				// then we should include it in the subring but continuing selecting instances.