--- a/ring/ring.go
+++ b/ring/ring.go
@@ -697,7 +697,8 @@ func (r *Ring) updateRingMetrics(compareResult CompareResult) {
 //
 // Subring returned by this method does not contain instances that have read-only field set.
 func (r *Ring) ShuffleShard(identifier string, size int) ReadRing {
-	// Use all instances if shuffle sharding is disabled, or it covers all instances anyway.
+	// Use all possible instances if shuffle sharding is disabled. We don't set size to r.InstancesCount(), because
+	// that could lead to wrong instances being returned when ring zones are unbalanced.
 	// Reason for not returning entire ring directly is that we need to filter out read-only instances.
 	if size <= 0 {
 		size = math.MaxInt