**What this PR does**:
PR #554 introduced a change to shuffle shard computation: when shard size <= 0 or size >= number of instances, returned shuffle shard can be different from shuffle shard before PR #554, if there are unbalanced zones in the ring.

For example, let's say there are three zones:
* zone-a, 3 instances
* zone-b, 2 instances
* zone-c, 1 instance

Before PR #554 shuffleShard with size = 0 or size = 6 would return entire ring.

After PR #554, shuffle shard with size = 0 or size = 6 returns ring with 5 instances, (2 from zone-a, 2 from zone-b, and 1 from zone-c).

This PR modifies that:
1. when shard size <= 0, shuffle shard now returns ring with all instances (that are not read-only)
2. when shard size == 6, shuffle shard now correctly returns only 2 instances from zone-a and zone-b, and 1 instance from zone-c.

Note that point 2 is still a change compared to before PR #554 was merged, **however previous version of the code was buggy**: if new instance is added in zone-c, and we request shard size 6 again, we will only get 6 instances back -- one instance from zone-a is now removed from the shard, and even though it has previously received data, it won't be queried anymore.

Since we now pass `math.MaxInt` as size to `shuffleShard` method, some changes were done to the method:
* map allocation hint is capped at "number of instances"

**Checklist**
- [x] Tests updated
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`