@pstibrany 2024-08-16
I've run shuffle shard benchmarks on `main`, this PR and this PR with optimization what stops iteration over tokens as soon as all instances in the zone were checked (see patch below).

<details><summary>Benchmark results</summary>
<p>
```console
goos: darwin
goarch: arm64
pkg: github.com/grafana/dskit/ring
                                                                                │   main.txt    │               pr566.txt               │         pr566-with-opts.txt          │
                                                                                │    sec/op     │    sec/op      vs base                │    sec/op      vs base               │
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_0-10             303.3µ ± ∞ ¹    650.4µ ± ∞ ¹  +114.44% (p=0.008 n=5)    298.1µ ± ∞ ¹        ~ (p=0.222 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-10             19.52µ ± ∞ ¹    19.40µ ± ∞ ¹    -0.58% (p=0.032 n=5)    19.89µ ± ∞ ¹        ~ (p=0.151 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-10            53.05µ ± ∞ ¹    53.07µ ± ∞ ¹         ~ (p=1.000 n=5)    54.44µ ± ∞ ¹   +2.62% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-10            162.3µ ± ∞ ¹    162.5µ ± ∞ ¹         ~ (p=1.000 n=5)    164.1µ ± ∞ ¹   +1.11% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_0-10             434.2µ ± ∞ ¹    665.1µ ± ∞ ¹   +53.16% (p=0.008 n=5)    337.5µ ± ∞ ¹  -22.28% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-10             45.40µ ± ∞ ¹    45.49µ ± ∞ ¹         ~ (p=0.690 n=5)    45.94µ ± ∞ ¹   +1.19% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-10            87.47µ ± ∞ ¹    86.94µ ± ∞ ¹         ~ (p=0.151 n=5)    88.08µ ± ∞ ¹   +0.70% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-10            201.1µ ± ∞ ¹    200.7µ ± ∞ ¹         ~ (p=0.690 n=5)    202.6µ ± ∞ ¹   +0.75% (p=0.016 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_0-10            659.3µ ± ∞ ¹   1372.2µ ± ∞ ¹  +108.11% (p=0.008 n=5)    658.3µ ± ∞ ¹        ~ (p=0.548 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-10            19.41µ ± ∞ ¹    19.40µ ± ∞ ¹         ~ (p=1.000 n=5)    20.22µ ± ∞ ¹   +4.17% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-10           53.46µ ± ∞ ¹    53.30µ ± ∞ ¹         ~ (p=0.095 n=5)    54.58µ ± ∞ ¹   +2.09% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-10           162.6µ ± ∞ ¹    162.3µ ± ∞ ¹         ~ (p=0.841 n=5)    163.9µ ± ∞ ¹   +0.79% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_0-10           1159.9µ ± ∞ ¹   1383.9µ ± ∞ ¹   +19.31% (p=0.008 n=5)    708.1µ ± ∞ ¹  -38.95% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-10            45.34µ ± ∞ ¹    45.41µ ± ∞ ¹         ~ (p=0.651 n=5)    46.27µ ± ∞ ¹        ~ (p=0.151 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-10           86.83µ ± ∞ ¹    87.03µ ± ∞ ¹         ~ (p=0.421 n=5)    88.17µ ± ∞ ¹   +1.54% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-10           200.3µ ± ∞ ¹    200.8µ ± ∞ ¹         ~ (p=0.548 n=5)    202.9µ ± ∞ ¹   +1.33% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_0-10           8.842m ± ∞ ¹   17.550m ± ∞ ¹   +98.47% (p=0.008 n=5)    8.780m ± ∞ ¹        ~ (p=1.000 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-10           17.27µ ± ∞ ¹    17.17µ ± ∞ ¹         ~ (p=0.548 n=5)    20.27µ ± ∞ ¹  +17.39% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-10          49.87µ ± ∞ ¹    49.66µ ± ∞ ¹         ~ (p=0.421 n=5)    53.26µ ± ∞ ¹   +6.79% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-10          157.3µ ± ∞ ¹    157.2µ ± ∞ ¹         ~ (p=1.000 n=5)    161.2µ ± ∞ ¹   +2.51% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_0-10          15.434m ± ∞ ¹   17.660m ± ∞ ¹   +14.43% (p=0.008 n=5)    9.288m ± ∞ ¹  -39.82% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-10           41.27µ ± ∞ ¹    41.17µ ± ∞ ¹         ~ (p=0.079 n=5)    44.27µ ± ∞ ¹   +7.26% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-10          80.06µ ± ∞ ¹    79.82µ ± ∞ ¹         ~ (p=0.095 n=5)    82.61µ ± ∞ ¹   +3.20% (p=0.008 n=5)
Ring_ShuffleShard/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-10          190.4µ ± ∞ ¹    189.9µ ± ∞ ¹         ~ (p=0.310 n=5)    194.0µ ± ∞ ¹   +1.87% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_0-10       1.594µ ± ∞ ¹    1.564µ ± ∞ ¹         ~ (p=0.095 n=5)    1.582µ ± ∞ ¹        ~ (p=0.341 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_3-10       144.0n ± ∞ ¹    140.0n ± ∞ ¹    -2.78% (p=0.016 n=5)    139.7n ± ∞ ¹   -2.99% (p=0.032 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_10-10      353.5n ± ∞ ¹    339.9n ± ∞ ¹    -3.85% (p=0.008 n=5)    346.8n ± ∞ ¹        ~ (p=0.310 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_1,_shard_size_=_30-10     1011.0n ± ∞ ¹    962.0n ± ∞ ¹    -4.85% (p=0.008 n=5)   1000.0n ± ∞ ¹        ~ (p=0.690 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_0-10       1.594µ ± ∞ ¹    1.578µ ± ∞ ¹         ~ (p=0.310 n=5)    1.590µ ± ∞ ¹        ~ (p=0.548 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_3-10       143.1n ± ∞ ¹    139.0n ± ∞ ¹         ~ (p=0.063 n=5)    138.6n ± ∞ ¹   -3.14% (p=0.032 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_10-10      404.1n ± ∞ ¹    395.1n ± ∞ ¹         ~ (p=0.222 n=5)    398.2n ± ∞ ¹        ~ (p=0.421 n=5)
Ring_ShuffleShardCached/num_instances_=_50,_num_zones_=_3,_shard_size_=_30-10      969.4n ± ∞ ¹    944.1n ± ∞ ¹         ~ (p=0.095 n=5)    951.1n ± ∞ ¹        ~ (p=0.548 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_0-10      3.461µ ± ∞ ¹    3.516µ ± ∞ ¹         ~ (p=0.421 n=5)    3.415µ ± ∞ ¹        ~ (p=0.095 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_3-10      144.5n ± ∞ ¹    141.5n ± ∞ ¹         ~ (p=0.548 n=5)    140.6n ± ∞ ¹        ~ (p=0.151 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_10-10     354.6n ± ∞ ¹    345.0n ± ∞ ¹         ~ (p=0.056 n=5)    348.6n ± ∞ ¹        ~ (p=0.222 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_1,_shard_size_=_30-10     988.3n ± ∞ ¹    963.1n ± ∞ ¹         ~ (p=0.421 n=5)    962.8n ± ∞ ¹        ~ (p=0.690 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_0-10      3.405µ ± ∞ ¹    3.369µ ± ∞ ¹         ~ (p=0.151 n=5)    3.398µ ± ∞ ¹        ~ (p=1.000 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_3-10      143.1n ± ∞ ¹    143.8n ± ∞ ¹         ~ (p=0.841 n=5)    141.5n ± ∞ ¹        ~ (p=0.421 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_10-10     421.8n ± ∞ ¹    402.5n ± ∞ ¹    -4.58% (p=0.008 n=5)    399.5n ± ∞ ¹   -5.29% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_100,_num_zones_=_3,_shard_size_=_30-10     985.4n ± ∞ ¹    975.5n ± ∞ ¹         ~ (p=0.651 n=5)    986.0n ± ∞ ¹        ~ (p=0.841 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_0-10     54.12µ ± ∞ ¹    53.98µ ± ∞ ¹         ~ (p=0.690 n=5)    53.58µ ± ∞ ¹        ~ (p=0.548 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_3-10     141.9n ± ∞ ¹    138.2n ± ∞ ¹    -2.61% (p=0.032 n=5)    138.7n ± ∞ ¹        ~ (p=0.095 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_10-10    350.9n ± ∞ ¹    350.0n ± ∞ ¹         ~ (p=0.690 n=5)    350.0n ± ∞ ¹        ~ (p=0.690 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_1,_shard_size_=_30-10   1008.0n ± ∞ ¹    949.9n ± ∞ ¹    -5.76% (p=0.032 n=5)    997.8n ± ∞ ¹        ~ (p=1.000 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_0-10     54.20µ ± ∞ ¹    55.17µ ± ∞ ¹         ~ (p=0.151 n=5)    54.27µ ± ∞ ¹        ~ (p=0.310 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_3-10     140.9n ± ∞ ¹    138.0n ± ∞ ¹         ~ (p=0.056 n=5)    137.6n ± ∞ ¹   -2.34% (p=0.008 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_10-10    401.5n ± ∞ ¹    401.6n ± ∞ ¹         ~ (p=0.841 n=5)    409.9n ± ∞ ¹   +2.09% (p=0.024 n=5)
Ring_ShuffleShardCached/num_instances_=_1000,_num_zones_=_3,_shard_size_=_30-10   1001.0n ± ∞ ¹    994.8n ± ∞ ¹         ~ (p=0.802 n=5)   1026.0n ± ∞ ¹        ~ (p=0.222 n=5)
Ring_ShuffleShard_512Tokens-10                                                     164.4µ ± ∞ ¹    164.4µ ± ∞ ¹         ~ (p=1.000 n=5)    172.0µ ± ∞ ¹   +4.67% (p=0.008 n=5)
Ring_ShuffleShard_LargeShardSize-10                                                7.819m ± ∞ ¹    7.788m ± ∞ ¹         ~ (p=0.690 n=5)    7.815m ± ∞ ¹        ~ (p=0.690 n=5)
Ring_ShuffleShard_ShardSize_0-10                                                   2.199m ± ∞ ¹    4.667m ± ∞ ¹  +112.20% (p=0.008 n=5)    2.205m ± ∞ ¹        ~ (p=0.310 n=5)
geomean                                                                            14.42µ          15.35µ          +6.46%                  14.15µ         -1.88%
¹ need >= 6 samples for confidence interval at level 0.95
```
</p>
</details> 

We can see that all `ShuffleShard` calls with shardSize=0 are now slower, as expected, and optimized version of the code can make those cases faster. However when caching is used, benchmarks show no slowdown at all even without optimization. And optimization makes other cases bit slower and code more complex.

Given that we use caching on our write path in Mimir, I will proceed with merging the code as-is, without the optimization.

<details><summary>ring.go optimization (patch)</summary>
<p>

```patch
Index: ring/ring.go
IDEA additional info:
Subsystem: com.intellij.openapi.diff.impl.patch.CharsetEP
<+>UTF-8
===================================================================
diff --git a/ring/ring.go b/ring/ring.go
--- a/ring/ring.go	(revision 35a77d0ac010034cc99fb782fe898183e73ad549)
+++ b/ring/ring.go	(date 1723818983954)
@@ -775,14 +775,21 @@
 	// We need to iterate zones always in the same order to guarantee stability.
 	for _, zone := range actualZones {
 		var tokens []uint32
+		var maxCheckedInstances int
 
 		if r.cfg.ZoneAwarenessEnabled {
 			tokens = r.ringTokensByZone[zone]
+			maxCheckedInstances = r.instancesCountPerZone[zone]
 		} else {
 			// When zone-awareness is disabled, we just iterate over 1 single fake zone
 			// and use all tokens in the ring.
 			tokens = r.ringTokens
+			maxCheckedInstances = len(r.ringDesc.Ingesters)
 		}
+
+		// If we checked all maxCheckedInstances, we can stop searching for more instances per zone, since there can't be any additional ones.
+		// We cannot use 'shard' map for this check, because not all instances are stored into that map. (Only those included in the shard)
+		checkedInstances := make(map[string]struct{}, maxCheckedInstances)
 
 		// Initialise the random generator used to select instances in the ring.
 		// Since we consider each zone like an independent ring, we have to use dedicated
@@ -793,7 +800,7 @@
 		// To select one more instance while guaranteeing the "consistency" property,
 		// we do pick a random value from the generator and resolve uniqueness collisions
 		// (if any) continuing walking the ring.
-		for i := 0; i < numInstancesPerZone; i++ {
+		for i := 0; i < numInstancesPerZone && len(checkedInstances) < maxCheckedInstances; i++ {
 			start := searchToken(tokens, random.Uint32())
 			iterations := 0
 			found := false
@@ -815,6 +822,8 @@
 					continue
 				}
 
+				checkedInstances[info.InstanceID] = struct{}{}
+
 				instanceID := info.InstanceID
 				instance := r.ringDesc.Ingesters[instanceID]
 
```


</p>
</details>